epic_events/requirements.txt
click==8.1.7
colorama==0.4.6
peewee==3.17.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
shellingham==1.5.4
typer==0.9.0
typing_extensions==4.8.0
